Refactor Queue entry to be closer with QueueEntry
